==> Notebook_1_dataset_efetividade.ipynb <==
## Notebook_1 - Coleta de dados - Dataset Efetividade + API

Objetivo: Formar uma dataset com os dados da efetividade (passageiros que foram fiscalizados e resultado) com os dados do API (informações do voo e dos passageiros no voo). 
Os dados da efetividade foram retirados do Sistema e-DBV e os dados API foram obtidos pelo Datalake interno da Receita Federal. 

import pandas as pd
import numpy as np
pd.set_option('display.max_columns',None)

Juntando os arquivos baixados da efetividade

gig1 = pd.read_csv('gig1.csv',sep=';')
gig2 = pd.read_csv('gig2.csv',sep=';')
gig3 = pd.read_csv('gig3.csv',sep=';')
gig4 = pd.read_csv('gig4.csv',sep=';')
gig5 = pd.read_csv('gig5.csv',sep=';')
gig6 = pd.read_csv('gig6.csv',sep=';')

gru1 = pd.read_csv('gru1.csv',sep=';')
gru2 = pd.read_csv('gru2.csv',sep=';')
gru3 = pd.read_csv('gru3.csv',sep=';')
gru4 = pd.read_csv('gru4.csv',sep=';')
gru5 = pd.read_csv('gru5.csv',sep=';')
gru6 = pd.read_csv('gru6.csv',sep=';')
gru7 = pd.read_csv('gru7.csv',sep=';')
gru8 = pd.read_csv('gru8.csv',sep=';')
gru9 = pd.read_csv('gru9.csv',sep=';')
gru10 = pd.read_csv('gru10.csv',sep=';')
gru11 = pd.read_csv('gru11.csv',sep=';')
gru12 = pd.read_csv('gru12.csv',sep=';')
gru13 = pd.read_csv('gru13.csv',sep=';')
gru14 = pd.read_csv('gru14.csv',sep=';')
gru15 = pd.read_csv('gru15.csv',sep=';')
gru16 = pd.read_csv('gru16.csv',sep=';')
gru17 = pd.read_csv('gru17.csv',sep=';')
gru18 = pd.read_csv('gru18.csv',sep=';')
gru19 = pd.read_csv('gru19.csv',sep=';')
gru20 = pd.read_csv('gru20.csv',sep=';')
gru21 = pd.read_csv('gru21.csv',sep=';')
gru22 = pd.read_csv('gru22.csv',sep=';')
gru23 = pd.read_csv('gru23.csv',sep=';')
gru24 = pd.read_csv('gru24.csv',sep=';')



DF_EFETIVIDADE 

frames = [gig1,gig2,gig3,gig4,gig5,gig6,gru1,gru2,gru3,gru4,gru5,gru6,gru7,gru8,gru9,gru10,gru11,gru12,gru13,gru14,gru15,gru16,gru17,gru18,gru19,gru20,gru21,gru22,gru23,gru24]
df = pd.concat(frames)
# df.head(2)

df.info()

Retirando colunas que não serão usadas e não irão interferir na análise ou cruzamento de dados

df_efetividade = df.copy()
df_efetividade = df_efetividade.drop(['Número da Fiscalização'], axis=1)
df_efetividade = df_efetividade.drop(df_efetividade.columns[[0,74,73,72,71,70,69,68,67,66,65,64,63,62,61,60,59,58,57,56,55,54]], axis=1)

df_efetividade['Recinto'] = df_efetividade['Recinto'].replace({'São Paulo/Guarulhos (GRU)':'GRU'})
df_efetividade['Recinto'] = df_efetividade['Recinto'].replace({'Rio de Janeiro/Galeão (GIG)':'GIG'})
#df

Criando um atributo para merge. Evitar repetição. VU + data

df_efetividade['vu chegada'] = df_efetividade['Data de Chegada do Voo'].astype(str).apply(lambda x: x[0:10].replace('/',''))
df_efetividade['vu chegada'] = df_efetividade['Código do VU'].astype(str) + df_efetividade['vu chegada']

df_efetividade['Data de Chegada do Voo2'] = pd.to_datetime(df_efetividade['Data de Chegada do Voo'],format='%d/%m/%Y', errors='ignore') #evitar erros de troca de dia por mês.
df_efetividade['Data de Chegada do Voo2'] = pd.to_datetime(df_efetividade['Data de Chegada do Voo2']) 
df_efetividade['data chegada sem hora'] = df_efetividade['Data de Chegada do Voo2'].dt.strftime('%D')

# df_efetividade[['Data de Chegada do Voo','Data de Chegada do Voo2','data chegada sem hora','vu chegada']].head(4)

Verificando os ficalizados que tiveram ocorrência sem espontaneidade

df_efetividade['fiscalizados_rapida'] = np.where(df_efetividade['Fisc. Rápida - Responsável Início - CPF'].isna(), 0, 1) 
df_efetividade['fiscalizados_detalhada'] = np.where(df_efetividade['Fisc. Detalhada - Responsável Término - CPF'].isna(), 0, 1)
df_efetividade['fiscalizados'] = df_efetividade['fiscalizados_rapida']+df_efetividade['fiscalizados_detalhada']
df_efetividade['fiscalizados'] = np.where(df_efetividade['fiscalizados'] > 0,1,0)

df_efetividade['ocorrencia'] = np.where(df_efetividade['Tipo'] == 'Ocorrência', 1, 0)
df_efetividade['nao_espontaneo'] = np.where(df_efetividade['Ocorrência com espontaneidade?'] == 'Não', 1, 0)
df_efetividade['resultado'] = df_efetividade['ocorrencia'] + df_efetividade['nao_espontaneo']
df_efetividade['resultado'] = np.where(df_efetividade['resultado'] > 1,1,0)



df_efetividade = df_efetividade.drop(columns=['Data da Pontuação', 'Tipo da Modulação', 'Situação do Processamento', 'Pontuação em Escopos Nacionais', 'Pontuação em Escopos Locais', 'Data da Seleção', 'Tipo da Seleção', 'Abrangência', 'Responsável pela Seleção Manual - CPF', 'Responsável pela Seleção Manual - Nome', 'Data do Carregamento', 'Data da Identificação', 'Evento de Identificação', 'Número da Declaração', 'Resultado da Análise de Risco da Declaração (canal)', 'Total de Porte de Valores Declarado (R$)', 'Total de Bens Declarado (US$)', 'Responsável pelo Registro - CPF', 'Responsável pelo Registro - Nome', 'Identificação por documento - Responsável CPF', 'Identificação por documento - Responsável Nome', 'Fisc. Rápida - Data do Início', 'Fisc. Rápida - Responsável Início - CPF', 'Fisc. Rápida - Responsável Início - Nome', 'Fisc. Rápida - Data do Término', 'Fisc. Rápida - Responsável Término - CPF', 'Fisc. Rápida - Responsável Término - Nome', 'Fisc. Rápida - Origem da Informação', 'Fisc. Detalhada - Número da Fiscalização', 'Fisc. Detalhada - Data do Início', 'Fisc. Detalhada - Responsável Início - CPF', 'Fisc. Detalhada - Responsável Início - Nome', 'Fisc. Detalhada - Data do Término', 'Fisc. Detalhada - Tipo do Término', 'Fisc. Detalhada - Responsável Término - CPF', 'Fisc. Detalhada - Responsável Término - Nome', 'Fisc. Detalhada - Origem da Informação', 'Número eDBV', 'Tipo', 'Data do último termo', 'Declaração Falsa ou Inexata?', 'Ocorrência com espontaneidade?','fiscalizados_rapida','fiscalizados_detalhada','ocorrencia','nao_espontaneo','CPF'])

Retirando as linhas duplicadas e somando a indicação dos fiscalizados e resultado

df_efetividade = df_efetividade.groupby(['Recinto','Cia/Número do Voo', 'Aeroporto de Origem do Voo','Código do VU','Número do Documento principal','País Emissor do Documento principal','Nome principal','Nacionalidade','Data de Nascimento','data chegada sem hora', 'vu chegada'],as_index=False, dropna=False)['fiscalizados', 'resultado'].sum()
# df_efetividade.sort_values('fiscalizados').tail(10)

Transformando os dados em fiscalizados e resultado em binários

df_efetividade['fiscalizados_0/1'] = np.where(df_efetividade['fiscalizados'] > 0, 1, 0)
df_efetividade['resultado_0/1'] = np.where(df_efetividade['resultado'] > 0, 1, 0)
# df_efetividade.sort_values('resultado').tail(10)

df_efetividade = df_efetividade.drop(columns=['fiscalizados','resultado'])

Retirando passageiros sem voo. Ex: declarante voo de saída

df_efetividade = df_efetividade.loc[df_efetividade['Cia/Número do Voo'].notna()]
# df_efetividade

--------------------------------------------------------------------------------------------------------------------------------
Dados API 


api = pd.read_csv('API TABELA VOOS.csv', sep=';', encoding = 'latin1')
df_api = api.copy()
df_api.info()

Criando código para o merge 

df_api['vu chegada'] = df_api['Data da Chegada do Voo'].astype(str).apply(lambda x: x[0:10].replace('/',''))
df_api['vu chegada'] = df_api['Numero do Viajante Unico'].astype(str) + df_api['vu chegada']
df_api['Data da Chegada do Voo2'] = df_api['Data da Chegada do Voo'].astype(str).apply(lambda x: x[0:10].replace('/',''))

# df_api[['vu chegada','Data da Chegada do Voo','Data da Chegada do Voo2']].head(2)

Voos duplicadas. Mesma situação dos dados da fiscalização que são oriundos desses dados. 

df_api = df_api.drop(columns=['Data da Chegada do Voo'])
df_api = df_api.drop_duplicates() # retirando as linhas duplicadas
# df_api.shape

Repetidos

df_repetidos_api = df_api.copy()
df_repetidos_api = df_repetidos_api.drop_duplicates()
df_repetidos_api = df_repetidos_api.loc[df_repetidos_api['Tipo de Viajante'] != 'Tripulante'] # serão retirados após o merge com o df principal
df_repetidos_api = df_repetidos_api.loc[df_repetidos_api['Tipo de Viajante'] != 'Passageiro em trânsito'] # serão retirados após com o df principal
df_repetidos_api['Quantidade de Registros'] = 1
df_repetidos_api['qtt_pax_repetidos'] = df_repetidos_api.groupby('vu chegada')['Quantidade de Registros'].transform(np.sum)
# df_repetidos_api.loc[df_repetidos_api['qtt_pax_repetidos']>1].sort_values('vu chegada')

-----------------------------------
Junção da Efetividade (df) com os dados API (df_API)

df_efetividade.shape,df_api.shape

df_efetividade_api = pd.merge(df_efetividade,df_api, left_on='vu chegada',right_on='vu chegada', how='left')
df_efetividade_api.shape

Retirando Tripulantes e Passageiros em trânsito

df_efetivadade_api_pax = df_efetividade_api.loc[(df_efetividade_api['Tipo de Viajante'] != 'Tripulante') & (df_efetividade_api['Tipo de Viajante'] != 'Passageiro em trânsito')]


Número de passageiros com a mesma reserva

df_efetivadade_api_pax['Quantidade de Registros'] = 1
df_efetivadade_api_pax['qtd_pax_reserva'] = df_efetivadade_api_pax.groupby('Codigo Localizador da Reserva')['Quantidade de Registros'].transform(np.sum)

DataFrame Reserva x Resultado

df_reservas_resultado = df_efetivadade_api_pax[['resultado_0/1','Codigo Localizador da Reserva']]
df_reservas_resultado.columns = ['resultado_reserva', 'Codigo Localizador da Reserva']
df_reservas_resultado = df_reservas_resultado.loc[df_reservas_resultado['resultado_reserva'] > 0]
# df_reservas_resultado

Agrupamento. Retirando as reservas duplicadas, somando resultado e atribuinto 0 ou 1 para reservas com resultado.

df_reservas_resultado = df_reservas_resultado.groupby(['Codigo Localizador da Reserva'],as_index=False, dropna=False)['resultado_reserva'].sum() # agrupando
df_reservas_resultado['resultado_reserva'] = np.where(df_reservas_resultado['resultado_reserva'] > 0, 1, 0)
df_reservas_resultado = df_reservas_resultado.drop_duplicates()
# df_reservas_resultado.sort_values('resultado_reserva')

--------------------------------------------------------------------------------------------------------------------
Merge: df_efetividade_api_pax x df_reservas_fiscalizadas

print(df_efetivadade_api_pax.shape)
print(df_reservas_resultado.shape)

df_efetivadade_api_pax_reserva = pd.merge(df_efetivadade_api_pax,df_reservas_resultado,left_on='Codigo Localizador da Reserva',
                                        right_on='Codigo Localizador da Reserva', how='left')
df_efetivadade_api_pax_reserva['resultado_reserva'] = np.where(df_efetivadade_api_pax_reserva['resultado_reserva'].isna(),0,1)
df_efetivadade_api_pax_reserva['fiscalizados_reserva'] = np.where((df_efetivadade_api_pax_reserva['fiscalizados_0/1'] == 0) 
                                                                & (df_efetivadade_api_pax_reserva['resultado_reserva'] == 1),1,
                                                                df_efetivadade_api_pax_reserva['fiscalizados_0/1']) 

df_efetivadade_api_pax_reserva.isna().sum()

Desconsiderando os passageiros fiscalizados que os voos não foram identificados no dados API disponíveis.  

df_efetivadade_api_pax_reserva['fiscalizados_reserva'] = np.where(df_efetivadade_api_pax_reserva['Codigo IATA do destino do Voo'].isna(),0,df_efetivadade_api_pax_reserva['fiscalizados_reserva'])


Considerando passageiros sem código de reserva como viajando sozinho. qtt_pax_reserva = 1

df_efetivadade_api_pax_reserva.loc[(df_efetivadade_api_pax_reserva['Codigo Localizador da Reserva'].isna()) & (df_efetivadade_api_pax_reserva['Codigo IATA do local de desembarque do viajante'].notna())].head(5)

df_efetivadade_api_pax_reserva['qtd_pax_reserva'] = np.where(df_efetivadade_api_pax_reserva['qtd_pax_reserva'].isna(),1,df_efetivadade_api_pax_reserva['qtd_pax_reserva'])
df_efetivadade_api_pax_reserva.loc[(df_efetivadade_api_pax_reserva['Codigo Localizador da Reserva'].isna()) & (df_efetivadade_api_pax_reserva['Codigo IATA do local de desembarque do viajante'].notna())].head()

---------------------------------------------------------------------------------------------------------------------------------------------------------
---------------------------------------------------------------------------------------------------------------------------------------------------------
PASSAGEIROS FISCALIZADOS E RESULTADO 

df_fiscalizados_resultado = df_efetivadade_api_pax_reserva.loc[df_efetivadade_api_pax_reserva['fiscalizados_reserva']>0]
# df_fiscalizados_resultado

df_fiscalizados_resultado.columns

df_fiscalizados_resultado.isna().sum()

df_fiscalizados_resultado = df_fiscalizados_resultado.drop(columns=['fiscalizados_reserva','fiscalizados_0/1','resultado_0/1','Numero do Viajante Unico','Quantidade de Registros','Tipo de Viajante'])
# df_fiscalizados_resultado

# df_fiscalizados_resultado.to_csv('fiscalizados_resultado.csv',sep=';')

df_fiscalizados_resultado.shape
