==> notebook_2_fiscalizados_resultados_vu data.ipynb <==
## Notebook_2
Dataset fiscalizados_resultados_vu

Junção do dataset dos fiscalizados com resultado (notebook_1) e as informações de histórico de viagens, renda, ocorrências, tributação, retenção etc.
Os dados foram obtidos por meio de um script que acessa as informações via requisições web e leitura de json.

import pandas as pd
import numpy as np
pd.set_option('display.max_columns',None)

vu1 = pd.read_csv('VU1.csv',sep=';',encoding = 'latin1',dtype={'vu_cod':'category'})
vu2 = pd.read_csv('VU2.csv',sep=';',encoding = 'latin1',dtype={'vu_cod':'category'})
vu3 = pd.read_csv('VU3.csv',sep=';',encoding = 'latin1',dtype={'vu_cod':'category'})
vu4 = pd.read_csv('VU4.csv',sep=';',encoding = 'latin1',dtype={'vu_cod':'category'})
vu5 = pd.read_csv('VU5.csv',sep=';',encoding = 'latin1',dtype={'vu_cod':'category'})
frames = [vu1,vu2,vu3,vu4,vu5]
df_vu = pd.concat(frames)
df_vu = df_vu.drop_duplicates()
df_vu.info()


-------------------------------------------
Dataset dos fiscalizados com resultado para merge (notebook_1)

df_fiscalizados_resultado = pd.read_csv('fiscalizados_resultado.csv',sep=';',dtype={'Código do VU':'category'})
# df_fiscalizados_resultado

-----------------------------------------------
Merge

df_fisc_resul_vu = pd.merge(df_fiscalizados_resultado,df_vu, left_on='Código do VU',right_on='vu_cod', how='left')
df_fisc_resul_vu.info()

Estudo dos dados não encontrados

df_fisc_resul_vu['vu_cod'].isna().sum()

vu_faltando = df_fisc_resul_vu.loc[df_fisc_resul_vu['vu_cod'].isna()]


vu_faltando = vu_faltando.drop(columns=['Unnamed: 0'])
# vu_faltando.head(2)

# vu_faltando.loc[vu_faltando['resultado_reserva']>0]

# df_fisc_resul_vu[['Nome principal','nome_vu']].head(36700)

df_fisc_resul_vu.to_csv('fiscalizados_resultados_vu.csv',sep=';')
