==> notebook_4_fiscalizados_resutlados_vu_pnr_data_formapgto.ipynb <==
## Notebook_4 - Coleta de dados 

import pandas as pd
import numpy as np
pd.set_option('display.max_columns',None)

---------------------------
Dataset fiscalizados_resultados_VU para merge com os dados da reserva - notebook_2

df_fisc_res_vu = pd.read_csv('fiscalizados_resultados_vu.csv',sep=';')
df_fisc_res_vu = df_fisc_res_vu.drop(columns=['Unnamed: 0','Unnamed: 0.1'])
df_fisc_res_vu.info()

df_fisc_res_vu = df_fisc_res_vu.drop(['monitoramento_titulo_vu','monitoramento_situacao','infoInteresseTitulo_vu','infoInteresseDescricao_vu','ar_selecionado_vu','ar_justificativa_vu','ar_justificativaSelecao_vu',
                                    'tipoVoo_vu','ciaAerea_vu','sequencial_vu','numeroVoo_vu','aeroportoOrigem_vu','aeroportoDestino_vu','dataPartida_vu','dataChegada_vu','qtBagagens_vu','localizadorReserva_vu',
                                    'qtViajantesReserva_vu','tipoViajante_vu','nomeUtilizado_vu','nomeAeroportoOrigem_vu','nomeAeroportoDestino_vu','aeroportoOrigemViajante_vu','aeroportoOrigemLatitude_vu',
                                    'aeroportoOrigemLongitude_vu','aeroportoOrigemCidade_vu','aeroportoDestinoViajante_vu','aeroportoDestinoLatitude_vu','aeroportoDestinoLongitude_vu','aeroportoDestinoCidade_vu',
                                    'nomeAeroportoOrigemViajante_vu','nomeAeroportoDestinoViajante_vu','dataNascimento_vu-2','documentos_vu','apvoNrSq_vu','apvaNrSq_vu','apvjNrSq_vu','assento_vu','duracaoDias_vu','apvjQtPesoBagKg_vu',
                                    'qtdBagagemPesoSaida_vu','qtdBagagemPesoChegada_vu','dataMovimentoNoBrasil_vu','localizadorReservaComQtdViajantes_vu','quantidadePesoBagagem_vu','viagemAnterior.tipoVoo_vu','viagemAnterior.ciaAerea_vu',
                                    'viagemAnterior.sequencial_vu','viagemAnterior.numeroVoo_vu','viagemAnterior.aeroportoOrigem_vu','viagemAnterior.aeroportoDestino_vu','viagemAnterior.dataPartida_vu','viagemAnterior.dataChegada_vu',
                                    'viagemAnterior.qtBagagens_vu','viagemAnterior.localizadorReserva_vu','viagemAnterior.qtViajantesReserva_vu','viagemAnterior.tipoViajante_vu','viagemAnterior.nomeUtilizado_vu','viagemAnterior.nomeAeroportoOrigem_vu',
                                    'viagemAnterior.nomeAeroportoDestino_vu','viagemAnterior.aeroportoOrigemViajante_vu','viagemAnterior.aeroportoOrigemLatitude_vu','viagemAnterior.aeroportoOrigemLongitude_vu','viagemAnterior.aeroportoOrigemCidade_vu',
                                    'viagemAnterior.aeroportoDestinoViajante_vu','viagemAnterior.aeroportoDestinoLatitude_vu','viagemAnterior.aeroportoDestinoLongitude_vu','viagemAnterior.aeroportoDestinoCidade_vu','viagemAnterior.nomeAeroportoOrigemViajante_vu',
                                    'viagemAnterior.nomeAeroportoDestinoViajante_vu','viagemAnterior.dataNascimento_vu','viagemAnterior.documentos_vu','viagemAnterior.apvoNrSq_vu','viagemAnterior.apvaNrSq_vu','viagemAnterior.apvjNrSq_vu',
                                    'viagemAnterior.assento_vu','viagemAnterior.duracaoDias_vu','viagemAnterior.apvjQtPesoBagKg_vu','viagemAnterior.viagemAnterior	_vu','viagemAnterior.qtdBagagemPesoSaida	_vu','viagemAnterior.qtdBagagemPesoChegada	_vu',
                                    'viagemAnterior.dataMovimentoNoBrasil	_vu','viagemAnterior.quantidadePesoBagagem_vu','viagemEntrada30Dias_vu',
                                    'documento_num_vu','documento_tipo_vu','documento_siglaEmissor_vu','cod_join_vu','link_vu'],axis=1) #'viagemAnterior.localizadorReservaComQtdViajantes _vua
# df_fisc_res_vu.head(2)

--------------------------------------------------------------------------------------------------------------------------------
PNR - Dados da reserva de passageiro estruturados por meio de script 

pnr1 = pd.read_csv('PNR_1.csv',sep=';',encoding = 'latin1')
pnr2 = pd.read_csv('PNR_2.csv',sep=';',encoding = 'latin1')
pnr3 = pd.read_csv('PNR_3.csv',sep=';',encoding = 'latin1')
pnr4 = pd.read_csv('PNR_4.csv',sep=';',encoding = 'latin1')
pnr5 = pd.read_csv('PNR_5.csv',sep=';',encoding = 'latin1')
pnr6 = pd.read_csv('PNR_6.csv',sep=';',encoding = 'latin1')
pnr7 = pd.read_csv('PNR_7.csv',sep=';',encoding = 'latin1')
pnr8 = pd.read_csv('PNR_8.csv',sep=';',encoding = 'latin1')
pnr9 = pd.read_csv('PNR_9.csv',sep=';',encoding = 'latin1')
pnr10 = pd.read_csv('PNR_10.csv',sep=';',encoding = 'latin1')
pnr11 = pd.read_csv('PNR_11.csv',sep=';',encoding = 'latin1')
frames = [pnr1,pnr2,pnr3,pnr4,pnr5,pnr6,pnr7,pnr8,pnr9,pnr10,pnr11]
df_pnr = pd.concat(frames)
df_pnr = df_pnr.drop_duplicates()
df_pnr.info()


# df_pnr.columns

Forma de pagamento 

df_pnr2 = df_pnr.copy()
df_pnr2 = df_pnr2[['codigo_reserva_pnr','voo_origem_pnr','data_reserva_pnr','forma_pgto_pnr']]

df_pnr2.shape

## Retirando Aeroportos brasileiros de origem

df_pnr_nacional = df_pnr2['voo_origem_pnr'].sort_values()
df_pnr_nacional = df_pnr_nacional.drop_duplicates()
#df_pnr_nacional.head(70)

df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'AJU']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'BEL']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'BPS']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'BSB']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'CGH']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'CGR']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'CNF']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'CWB']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'FLN']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'FOR']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'GIG']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'GRU']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'GYN']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'IGU']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'IMP']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'IOS']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'JJG']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'JPA']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'MAO']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'MCZ']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'MGF']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'NAT']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'NVT']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'PMW']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'POA']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'PVH']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'REC']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'CGB']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'LDB']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'RBR']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'SJK']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'SJP']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'SLZ']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'UDI']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'VIX']
df_pnr2 = df_pnr2.loc[df_pnr2['voo_origem_pnr'] != 'SDU']



## Identificando as reservas e a data da reserva

df_pnr2 = df_pnr2[['codigo_reserva_pnr','data_reserva_pnr','forma_pgto_pnr']]
df_pnr2 = df_pnr2.drop_duplicates()  
# df_pnr2.sort_values('codigo_reserva_pnr').head(5)

df_pnr2.shape

-------------------------------------------------------------------------------------------------------------------------------------
Merge. df_fisc_res_vu (fiscalizados com resultado) x df_pnr2 (dados da reserva)

df_fisc_res_vu.shape, df_pnr2.shape

df_principal = pd.merge(df_fisc_res_vu,df_pnr2, left_on='Codigo Localizador da Reserva',right_on='codigo_reserva_pnr', how='left')
# df_principal

Separando os viajantes que possuem dados PNR

df_principal_pnr = df_principal[(df_principal.codigo_reserva_pnr.notna())]
# df_principal_pnr 

Análise dos DFs

print('Dos passagerios fiscalizados\n----------------------------')
print('Sem considerar o PNR, número de passageiros: ',df_principal['Recinto'].count(),'\nCom resultado: ', df_principal['resultado_reserva'].sum(),'\n----------------------------')
print('No df com dados PNR, número de passageiros:', df_principal_pnr['codigo_reserva_pnr'].notna().sum(),'\nCom resultado: ', df_principal_pnr['resultado_reserva'].sum())

df_principal_pnr = df_principal_pnr.drop(columns=['data chegada sem hora','ocorrencia_tipo_24meses', '_ocorrecia_valores_ocorrencias_24meses_vu', 'oc_Apreensão Drogas_24meses_vu', 'oc_Apreensão Bens_24meses', 'oc_RTE não Desembaraçado_24meses', 'oc_RTE_24meses', 'oc_TECAT Fechado_24meses', 'oc_TECAT Aberto_24meses', 'oc_Mercadorias Retidas_24meses', 'oc_Moeda Validada_24meses', 'oc_Moeda Retida_24meses', 'oc_Liberado_24meses', 'bens_grupo_24meses', 'bens_valores_grupo_24meses', 'bens_Eletrônicos_24meses', 'bens_Bebidas Alcoólicas_24meses', 'bens_Artigos de Fumo_24meses', 'bens_Equipamento Médico e Odontológico_24meses', 'bens_Artigo Esportivo_24meses', 'bens_Instrumento Musical_24meses', 'bens_Acessório para Veículo_24meses', 'bens_Produto de Higiene e Beleza_24meses', 'bens_Vestuário_24meses', 'bens_Armas e Munições_24meses', 'bens_Brinquedo e Lazer_24meses', 'bens_Eletrodomésticos e Ferramentas_24meses', 'bens_Veículos, Parte e Peças para Veículos_24meses', 'bens_Origem Animal_24meses', 'bens_Origem Vegetal_24meses', 'bens_Porte de Valores_24meses', 'bens_Insumos Agropecuários_24meses', 'bens_Outros_24meses', 'bens_Entorpecentes_24meses',
                                'Recinto', 'Cia/Número do Voo', 'Data de Nascimento','vu_cod','vu chegada','versaoRegistro_vu','foto_possuiFoto_vu',
                                'contato_telefone_vu','contato_celular_vu','Codigo Localizador da Reserva',
                                'Número do Documento principal','Codigo IATA da origem do voo','analisesRisco_qt_vu',
                                'Nome principal','nome_vu','nome_cpf_vu','nomeResponsavel_vu','endereco_vu','endereco_vu','endereco_cep_vu','endereco_numero_vu','endereco_complemento_vu',
                                'dataNascimento_cpf_vu','codigo_reserva_pnr',
                                'selecao_qt_vu','fisc_rapida_qt_vu','fisc_detalhada_qt_vu','documento_emissao_vu','documento_emissao_vu','idade_vu',
                                'historicosContagem_vu','documentos_hist_vu','dirpf_anoExercicio_vu',
                                'dirpf_anoExercicio_vu','dirpf_hashDeclaracao_vu',
                                'dirpf_descricaoTipoContribuinte_vu','viagemAnterior.localizadorReservaComQtdViajantes\t_vu','cpf_cadastro_vu',
                                'endereco_bairro_vu','endereco_cdMunicipio_vu','endereco_uf_vu','endereco_nmMunicipio_vu',
                                'cnpj_razaoSocial_vu','cnpj_nomeFantasia_vu', 'cnpj_tipoParticipacao_vu','cnpj_capitalSocial_vu','cnpj_nmMunicipio_vu','cnpj_dataInicioFormatada_vu',
                                'cnpj_Formatado_vu','cnpj_atividadeEconomicaSecundaria_vu',
                                'doc_tipoDocumento_hist_vu','doc_dataValidade_hist_vu','dataPartidaHist_vu','localizadorReservaHist_vu',
                                'dirpf_totalRendimentoTributavelDependente_vu','dirpf_totalRendimentoNaoTributavelDependente_vu'
                                ])


df_principal_pnr.columns = df_principal_pnr.columns.str.lower().str.replace(' ', '_')

df_principal_pnr.dtypes

df_principal_pnr.to_csv('notebook_5_processamento_formapgto-novembro.csv', sep=';')

df_principal_pnr.shape
