==> notebook_5_tratamento_parte1.ipynb <==
## Notebook_5 Tratamento de dados - Parte 1 
Preparação e transformaçãos 

import pandas as pd
import numpy as np
from datetime import date, datetime
pd.set_option('display.max_columns',None)

def transf_string_inteiro_soma(coluna2):
    coluna = []
    for i in coluna2:
        soma = 0
        if type(i) == list:
            for x in i:
                x = int(x)
                soma += x
            coluna.append(soma)    
        else: 
            coluna.append(i)
    return coluna

def transf_string_inteiro(coluna2):
    coluna = []
    for i in coluna2:
        lista = []
        if type(i) == list:
            for x in i:
                x = int(x)
                lista.append(x)    
            coluna.append(lista)    
        else: 
            coluna.append(i)
    return coluna

def transformar_em_data_chegada(dt):
    if len(dt) > 7:
        dia, mes, ano = int(dt[0:2]), int(dt[2:4]), int(dt[4:])
        data = date(ano, mes, dia)
    if len(dt) < 8:
        dia, mes, ano = int(dt[0]), int(dt[1:3]), int(dt[3:])
        data = date(ano, mes, dia)
    return data

def transformar_em_data_chegada_merge(dt):
    dia, mes, ano = str(dt[0:2]), str(dt[3:5]), str(dt[6:10])
    data = dia + mes + ano
    return data

df_trat = pd.read_csv('df_principal_pnr_formapgto.csv',sep=';')
df_trat.columns = df_trat.columns.str.lower().str.replace(' ', '_').str.replace('código_do_vu', 'cod_vu')
df_trat = df_trat.drop(columns=['unnamed:_0'])
df_trat = df_trat.drop_duplicates()

df_trat.shape

------------------------
Código do VU

df_trat['cod_vu'] = df_trat['cod_vu'].astype(str)

Número do voo

df_trat['numero_do_voo'] = df_trat['numero_do_voo'].astype(int).astype(str)
df_trat['numero_do_voo'] = 'Z' + df_trat['numero_do_voo'].astype('str') # para ficar categorico

Tratando data de nascimento 

# df_trat.loc[df_trat['datanascimento_vu'].isna().sum()]
df_trat['datanascimento_vu'] = df_trat['datanascimento_vu'].fillna('29/03/1988')
df_trat['datanascimento_vu'].isna().sum()

def transformar_em_data_nascimento(dt):
    dia, mes, ano = int(dt[0:2]),int(dt[3:5]),int(dt[6:10])
    data = date(ano, mes, dia)
    return data

data_lista, data_lista2 = list(df_trat['datanascimento_vu']), []
for i in data_lista:
    nova_dt = transformar_em_data_nascimento(i)
    data_lista2.append(transformar_em_data_nascimento(i))
df_trat['datanascimento_vu'] = data_lista2


Tratando a data de chegada

data_lista, data_lista2 = list(df_trat['data_da_chegada_do_voo2'].astype(str)), []
for i in data_lista:
    data_lista2.append(transformar_em_data_chegada(i))

df_trat['data_da_chegada_do_voo'] = data_lista2

-----------------------------
FISCALIZAÇÃO
###### Verifica quantas fiscalizações anteriores no período foram realizadas antes do registro da fiscalização em questão

# Para comparação com os fiscalizados
df_trat['viagens_periodo_do_vu_trat'] = 1
df_trat['viagens_periodo_do_vu_trat'] = df_trat.groupby('cod_vu')['viagens_periodo_do_vu_trat'].transform(np.sum)
df_trat = df_trat.sort_values('data_da_chegada_do_voo')

df_trat = df_trat.sort_values('data_da_chegada_do_voo',ascending=False)
dic_contagem, lista_qtd = {} , []

for i in df_trat['cod_vu']:
    if i in dic_contagem:
        dic_contagem[i] += 1
        lista_qtd.append(dic_contagem[i])  
    else:
        lista_qtd.append(1)
        dic_contagem[i] = 1

df_trat['qtd_fisc_descendente_trat'] = lista_qtd
df_trat = df_trat.sort_values('data_da_chegada_do_voo')
df_trat['fiscalizacoes_qt_vu_trat'] = df_trat['fiscalizacoes_qt_vu'] - df_trat['qtd_fisc_descendente_trat']
df_trat['fiscalizacoes_qt_vu_trat'] = df_trat['fiscalizacoes_qt_vu_trat'].apply(lambda x: 0 if x<0 else x).fillna(13)
# df_trat[['cod_vu','qtd_fisc_descendente_trat','viagens_periodo_do_vu_trat',
#         'fiscalizacoes_qt_vu','fiscalizacoes_qt_vu_trat'
#         ]].loc[df_trat['viagens_periodo_do_vu_trat']>1].sort_values('cod_vu').head(4)

-----------------
VIAGENS - Chegada - API
###### Vericando o número de viagens na data de cada fiscalização a partir de todos os arquivos API do período

#arquivos do API - esse API é todos os voos, de entrada e de saída do país. 
api1 = pd.read_csv('VOOS jan mar.csv',sep=';', encoding = 'latin1')
api2 = pd.read_csv('VOOS abr jun.csv',sep=';', encoding = 'latin1')
df_api = pd.concat([api1,api2])
df_api.shape

Criando um código para o merge 

lista_data, lista_data_merge = list(df_api['Data da Chegada do Voo']), []
for i in lista_data:
    lista_data_merge.append(transformar_em_data_chegada_merge(i)) # transformando a data da chegada
df_api['Data da Chegada do Voo merge'] = lista_data_merge
df_api['vu chegada'] = df_api['Numero do Viajante Unico'].astype(str) + df_api['Data da Chegada do Voo merge']

# retirando repetidos (parte 1) por meio de um dicionário. Aqui, verifica se há repeditos para depois retirar esses repetidos com LOC
dic_contagem, lista_qtd  = {}, []
for i in df_api['vu chegada']:
    if i in dic_contagem:
        lista_qtd.append(dic_contagem[i])
        dic_contagem[i] += 1
    else:
        lista_qtd.append(0)
        dic_contagem[i] = 1

df_api['num_viagens_api'] = lista_qtd
# df_api[['Numero do Viajante Unico','num_viagens_api']].loc[df_api['num_viagens_api']>1].sort_values('Numero do Viajante Unico').head(2)

#Retirando repeditos (parte 2)
df_api = df_api.loc[df_api['num_viagens_api'] == 0] 

# somando o número de viagens para comparação
df_api['num_viagens_api'] = 1
df_api['num_viagens_api'] = df_api.groupby('Numero do Viajante Unico')['num_viagens_api'].transform(np.sum)
# df_api.loc[df_api['num_viagens_api']>1].sort_values('Numero do Viajante Unico').head(4)

# contado viagens anteriores à data da viagem atual.
df_api, dic_contagem, lista_qtd = df_api.sort_values('Data da Chegada do Voo',ascending=False), {}, []
for i in df_api['Numero do Viajante Unico']:
    if i in dic_contagem:
        dic_contagem[i] += 1
        lista_qtd.append(dic_contagem[i])
    else:
        lista_qtd.append(1)
        dic_contagem[i] = 1

df_api['num_viagens_api_subtrair'] = lista_qtd

df_api = df_api[['vu chegada','num_viagens_api_subtrair']]

-------------------------------------------------
Merge com o df_trat - número de viagens

lista_data = list(df_trat['data_da_chegada_do_voo'])
lista_dia, lista_mes, lista_ano = [], [], []
for i in lista_data:
    lista_dia.append(str(i.strftime('%d'))), lista_mes.append(str(i.strftime('%m'))), lista_ano.append(str(i.strftime('%Y')))

df_trat['data_merge_dia'], df_trat['data_merge_mes'], df_trat['data_merge_ano'] = lista_dia, lista_mes,lista_ano
df_trat['data_merge'] = df_trat['cod_vu'].astype(str) + df_trat['data_merge_dia'] + df_trat['data_merge_mes'] + df_trat['data_merge_ano']

df_trat.shape,df_api.shape

df_trat = pd.merge(df_trat,df_api, left_on='data_merge',right_on='vu chegada', how='left')
df_trat.shape

# Trando o número de viagens.
df_trat['qtviagens_vu_trat'] = df_trat['qtviagens_vu'] - df_trat['num_viagens_api_subtrair']
df_trat['qtviagens_vu_trat'] = df_trat['qtviagens_vu_trat'].fillna(219) #apenas um passageiros que foi consultado
df_trat[['qtviagens_vu','num_viagens_api_subtrair', 'qtviagens_vu_trat']].sample(2)

-----------------
OCORRÊNCIAS: calculando o número de ocorrência das informações do VU para quem deu resultado. 

df_trat = df_trat.sort_values('data_da_chegada_do_voo',ascending=False)
ls_vu, ls_resultado = list(df_trat['cod_vu']), list(df_trat['resultado_reserva']) 
dic_contagem, ls_hist_resultado = {}, []

index = 0
for vu in ls_vu: 
    if vu in dic_contagem:
        if ls_resultado[index] == 0:
            ls_hist_resultado.append(dic_contagem[vu])
        if ls_resultado[index] > 0:
            dic_contagem[vu] += 1
            ls_hist_resultado.append(dic_contagem[vu])
    if vu not in dic_contagem:
        if ls_resultado[index] == 0:
            dic_contagem[vu] = 0
            ls_hist_resultado.append(0)
        else:
            dic_contagem[vu] = 1
            ls_hist_resultado.append(1)
    index += 1
            
df_trat['declaracoes_qt_vu_trat'] = ls_hist_resultado
df_trat['declaracoes_qt_vu_trat'] = df_trat['declaracoes_qt_vu'] - df_trat['declaracoes_qt_vu_trat']
df_trat['declaracoes_qt_vu_trat'] = np.where(df_trat['declaracoes_qt_vu_trat'] <0, 0, df_trat['declaracoes_qt_vu_trat'])

df_trat['declaracoes_qt_vu_trat'] = df_trat['declaracoes_qt_vu_trat'].fillna(0) #apenas um passageiros que foi consultado

df_trat[['declaracoes_qt_vu','resultado_reserva','declaracoes_qt_vu_trat']].sort_values('resultado_reserva', ascending=False).head(4)


Documento - validade - ano - nan

df_trat['documento_validade_vu'] = pd.to_datetime(df_trat['documento_validade_vu'],format='%d/%m/%Y',errors='ignore')
df_trat['doc_ano_validade_trat'] = df_trat['documento_validade_vu'].dt.year
df_trat['doc_ano_validade_trat_2'] = df_trat['doc_ano_validade_trat'].loc[df_trat['doc_ano_validade_trat'].notna()]
df_trat['doc_ano_validade_trat'] = df_trat['documento_validade_vu'].fillna(df_trat['doc_ano_validade_trat_2'].median())
df_trat = df_trat.drop(columns=['doc_ano_validade_trat_2'])

Nacionalidade nan
###### identifiquei que os viajantes com nacionalidade faltante eram Brasileiros. 2

# df_trat.loc[df_trat['nacionalidade'].isna()]
df_trat['nacionalidade'] = df_trat['nacionalidade'].fillna('BRA') 

Dia da semana data de chegada


lista_data, lista_dia_semana = list(df_trat['data_da_chegada_do_voo']), []
for i in lista_data:
    lista_dia_semana.append(i.isoweekday())
df_trat['dia_semana_chegada_trat'] = lista_dia_semana

CPF e CPF do responsável

df_trat['cpf_trat'] = np.where(df_trat['cpf_vu'].notna(), True,df_trat['cpf_vu'])
df_trat['cpfresponsavel_vu'] = np.where(df_trat['cpfresponsavel_vu'].notna(), True,df_trat['cpfresponsavel_vu'])

Região Fiscal do CPF

df_trat['cpf_regiao_fiscal_trat'] = np.select([
                                    (df_trat['cpf_trat']==True)], 
                                    [(df_trat['cpf_vu'].str.slice(start=10,stop=11))], 
                                    default=df_trat['cpf_vu'])
                                    

-----------------
CNPJ

Quantidade de CNPJ's e estabelecimentos

df_trat['cnpj_qtd_trat'] = df_trat['cnpj_cod_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
df_trat['cnpj_qtd_trat'] = df_trat['cnpj_qtd_trat'].apply(lambda x: len(x) if type(x) == list else x)

df_trat['cnpj_nrestabelecimento_trat'] = df_trat['cnpj_nrestabelecimento_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
df_trat['cnpj_nrestabelecimento_trat'] = transf_string_inteiro_soma(df_trat['cnpj_nrestabelecimento_trat'])


Verificando a situação dos CPNJs

df_trat['situacao_cnpj_ativa_trat'] = df_trat['cnpj_situacao_vu'].str.contains('ATIVA')
df_trat['situacao_cnpj_ativa_trat'] = np.where(df_trat['situacao_cnpj_ativa_trat'] == True, 1, 0)
df_trat['situacao_cnpj_ativa_trat'] = np.select([(df_trat['cnpj_situacao_vu'].isna())], [(df_trat['cnpj_situacao_vu'])] , default=df_trat['situacao_cnpj_ativa_trat'])

df_trat['situacao_cnpj_baixada_trat'] = df_trat['cnpj_situacao_vu'].str.contains('BAIXADA')
df_trat['situacao_cnpj_baixada_trat'] = np.where(df_trat['situacao_cnpj_baixada_trat'] == True, 1, 0)
df_trat['situacao_cnpj_baixada_trat'] = np.select([(df_trat['cnpj_situacao_vu'].isna())], [(df_trat['cnpj_situacao_vu'])] , default=df_trat['situacao_cnpj_baixada_trat'])

df_trat['situacao_cnpj_inapta_trat'] = df_trat['cnpj_situacao_vu'].str.contains('INAPTA')
df_trat['situacao_cnpj_inapta_trat'] = np.where(df_trat['situacao_cnpj_inapta_trat'] == True, 1, 0)
df_trat['situacao_cnpj_inapta_trat'] = np.select([(df_trat['cnpj_situacao_vu'].isna())], [(df_trat['cnpj_situacao_vu'])], default=df_trat['situacao_cnpj_inapta_trat'])

df_trat['situacao_cnpj_nula_trat'] = df_trat['cnpj_situacao_vu'].str.contains('NULA')
df_trat['situacao_cnpj_nula_trat'] = np.where(df_trat['situacao_cnpj_nula_trat'] == True, 1, 0)
df_trat['situacao_cnpj_nula_trat'] = np.select([(df_trat['cnpj_situacao_vu'].isna())], [(df_trat['cnpj_situacao_vu'])], default=df_trat['situacao_cnpj_nula_trat'])

df_trat['situacao_cnpj_suspensa_trat'] = df_trat['cnpj_situacao_vu'].str.contains('SUSPENSA')
df_trat['situacao_cnpj_suspensa_trat'] = np.where(df_trat['situacao_cnpj_suspensa_trat'] == True, 1, 0)
df_trat['situacao_cnpj_suspensa_trat'] = np.select([(df_trat['cnpj_situacao_vu'].isna())], [(df_trat['cnpj_situacao_vu'])], default=df_trat['situacao_cnpj_suspensa_trat'])


---------------------
Gênero

#df_trat['genero_vu'].isna().sum()
#df_trat.loc[df_trat['genero_vu'].isna()]
df_trat['genero_vu'] = df_trat['genero_vu'].fillna('M')
df_trat['genero_vu'] = df_trat['genero_vu'].str.replace('U','F')

País_emissor_do_documento_principal


print(df_trat['país_emissor_do_documento_principal'].isna().sum())
# df_trat.loc[df_trat['país_emissor_do_documento_principal'].isna()]
df_trat.drop(df_trat[df_trat['país_emissor_do_documento_principal'].isna()].index, inplace=True)


EMISSOR DOCUMENTO

#df_trat['documento_emissor_vu'].isna().sum()
#df_trat.loc[df_trat['documento_emissor_vu'].isna()]
# df_trat.drop(df_trat[df_trat['cod_vu']== '20200107191840032'].index, inplace=True)
df_trat['documento_emissor_vu'] = df_trat['documento_emissor_vu'].fillna('CONGO, REPUBLICA DEMOCRATICA DO')
df_trat['documento_emissor_vu'].isna().sum()


Qtdd de documentos utilizados

print(df_trat['doc_numdocumento_hist_vu'].isna().sum())
df_trat['doc_qtd_trat'] = df_trat['doc_numdocumento_hist_vu'].str.replace('[','').str.replace(']','').str.split(',')
df_trat['doc_qtd_trat'] = df_trat['doc_qtd_trat'].apply(lambda x: len(x) if type(x) == list else x)
df_trat.drop(df_trat[df_trat['doc_qtd_trat'].isna()].index, inplace=True)

Idade na data da fiscalização 

df_trat['idade_trat'] = (df_trat['data_da_chegada_do_voo'] - df_trat['datanascimento_vu']).astype('timedelta64[Y]')


MONITORADO e INFORMAÇÃO DE INTERESSE

# O nan foram veritificados manualmente
print(df_trat['monitoramentos_qt_vu'].isna().sum())
# df_trat.loc[df_trat['monitoramentos_qt_vu'].isna()]
df_trat['monitoramentos_qt_vu'] = df_trat['monitoramentos_qt_vu'].fillna(0)

# Os nan foram veritificados manualmente
print(df_trat['infosinteresse_qt_vu'].isna().sum())
# df_trat.loc[df_trat['infosinteresse_qt_vu'].isna()]
df_trat['infosinteresse_qt_vu'] = df_trat['infosinteresse_qt_vu'].fillna(0)

DOCUMENTOS - N. DE PAÍSES EMISSORES

df_trat['doc_qtd__paises_trat'] = df_trat['doc_paisemissor_hist_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
df_trat['doc_qtd__paises_trat'] = df_trat['doc_qtd__paises_trat'].apply(lambda x: len(set(x)) if type(x) == list else x)
df_trat[['doc_paisemissor_hist_vu','doc_qtd__paises_trat']].head(5)


IATA DE ORIGEM - quantidade de vezes que o passageiro já originou de um determinado aeroporto (iata)

df_trat['origem_ls_trat'] = df_trat['aeroportoorigemviajantehist_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
df_trat['origem_ls_trat'] = df_trat['origem_ls_trat'].fillna('[]') 
df_trat['qtviagens_vu_trat'] = df_trat['qtviagens_vu_trat'].astype(int) 

ls_iatas_origem = list(df_trat['origem_ls_trat']) 
ls_qtviagens = list(df_trat['qtviagens_vu_trat'])
ls_iatas_fisc = []
index = 0 
for i in ls_iatas_origem:
    iatas_fiscalizacao = i[-(ls_qtviagens[index]):]
    ls_iatas_fisc.append(iatas_fiscalizacao)
    index +=1

df_trat['origem_ls_trat_2'] = ls_iatas_fisc

lista_iatas_origem = list(df_trat['origem_ls_trat_2']) 
lista_iata_origem = list(df_trat['codigo_iata_do_local_de_embarque_do_passageiro'])
lista_iata_origem_qtd = []
index = 0
for i in lista_iatas_origem:
    j = lista_iata_origem[index]
    lista_iata_origem_qtd.append(i.count(j))
    index +=1 
df_trat['qtd_iata_origem_trat'] = lista_iata_origem_qtd
df_trat[['origem_ls_trat','qtviagens_vu_trat','origem_ls_trat_2','codigo_iata_do_local_de_embarque_do_passageiro','qtd_iata_origem_trat']].sample(2)


QTD de países de origem

df_trat['qtd_paises_visitados_trat'] = df_trat['origem_qt_vu'].str.replace('[','').str.replace(']','').str.split(',') 
df_trat['qtd_paises_visitados_trat'] = df_trat['qtd_paises_visitados_trat'].apply(lambda x: len(x) if type(x) == list else x)


 QTD países de destino

df_trat['qtd_paises_destino_trat'] = df_trat['destino_qt_vu'].str.replace('[','').str.replace(']','').str.split(',') 
df_trat['qtd_paises_destino_trat'] = df_trat['qtd_paises_destino_trat'].apply(lambda x: len(x) if type(x) == list else x)


Número de cia utilizadas pelo viajante

# df_trat.loc[df_trat['ciaaereahist_vu'].isna()]
df_trat['ciaaereahist_vu'] = df_trat['ciaaereahist_vu'].fillna('[UA, CM, AD]')
# df_trat.loc[df_trat['ciaaereahist_vu'].isna()]

df_trat['qtd_ciaaerea_trat'] = df_trat['ciaaereahist_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
lista_1,lista_2 = list(df_trat['qtd_ciaaerea_trat']),[]
for i in lista_1:
    lista_3 = i
    lista_4 = []
    for j in lista_3:
        lista_4.append(j.strip(' '))
    lista_2.append(len(set(lista_4)))

df_trat['qtd_ciaaerea_trat'] = lista_2
df_trat[['ciaaereahist_vu','qtd_ciaaerea_trat']].sample(4)

Média de viajantes na mesma reserva
###### incluíndo todos os voos

df_trat['qtviajantesreservahist_vu2'] = df_trat['qtviajantesreservahist_vu'].str.replace('[','').str.replace(']','').str.replace(' ','').str.split(',')
df_trat['qtviajantesreservahist_vu2'].head(4)

df_trat['num_viagens_api_subtrair'] = df_trat['num_viagens_api_subtrair'].fillna(0)


df_trat['qtviajantesreservahist_vu2'] = df_trat['qtviajantesreservahist_vu2'].fillna('[0]')

lista1, lista2, lista_res  = list(df_trat['qtviajantesreservahist_vu2']), list(df_trat['num_viagens_api_subtrair']), []
index = 0
for i in lista1:
    j = int(lista2[index])
    lista_res.append(i[j:])
    index +=1

lista_res2 = []
for i in lista_res:
    lista_res3 = []
    for j in i:
        try:
            lista_res3.append(int(j))
        except:
            lista_res3.append(j)
    try:
        mean = sum(lista_res3)/len(lista_res3)
        lista_res2.append(mean)
    except:
        lista_res2.append(-1)

df_trat['qtviajantesreservahist_vu_media'] = lista_res2
df_trat[['qtviajantesreservahist_vu2','num_viagens_api_subtrair','qtviajantesreservahist_vu_media']].sample(5)


 RESERVA: Número de dias antes da chegada

def transformar_em_data_partida(dt):
    dia, mes, ano = int(dt[0:2]), int(dt[3:5]), int(dt[6:10])
    data = date(ano, mes, dia)
    return data

df_trat['data_reserva_pnr_trat'] = pd.to_datetime(df_trat['data_reserva_pnr'],format='%d/%m/%Y',errors='ignore')
df_trat['data_de_partida_do_voo_trat'] = df_trat['data_de_partida_do_voo'].apply(lambda x: x[0:10])

data_lista, data_lista2 = list(df_trat['data_de_partida_do_voo_trat']),[] # tratamento da data da partida
for i in data_lista:
    data_lista2.append(transformar_em_data_partida(i))
df_trat['data_de_partida_do_voo_trat'] = data_lista2

data_lista, data_lista2 = list(df_trat['data_reserva_pnr']), [] # tratamento da data da reserva
for i in data_lista:
    data_lista2.append(transformar_em_data_partida(i))
df_trat['data_reserva_pnr_trat'] = data_lista2

df_trat['reserva_dias_antes_partida_trat'] = df_trat['data_de_partida_do_voo_trat'] - df_trat['data_reserva_pnr_trat']
df_trat[['data_de_partida_do_voo','reserva_dias_antes_partida_trat','data_reserva_pnr']].sample(2)



 Cidade de embarque

df_trat['cidade_do_embarque_do_viajante'] = df_trat['cidade_do_embarque_do_viajante'].fillna('Najran') #apenas um nan foi identificado

 Padronização da nacionalidade, pais_emissor do documento principal, pais de origem do voo, pais de embarque do viajante, pais emissor do documento

df_trat['nacionalidade'] = df_trat['nacionalidade'].str.normalize('NFKD').str.encode('ascii', errors='ignore').str.decode('utf-8').str.replace(
                            ' ','').str.lower().str.replace(',','').str.replace('armeniarepublicada','amermenia').str.replace('chinarepublicapopular','china').str.replace(
                            'congorepublicademocraticado','congo')
df_trat['nacionalidade']=df_trat['nacionalidade'].str.replace('caboverderepublicade','caboverde').str.replace('estoniarepublicada','estonia').str.replace(
                            'georgiarepublicada','georgia').str.replace('croacia(republicada)','croacia').str.replace('hungriarepublicada','hungria').str.replace(
                            'irarepublicaislamicado','ira').str.replace('cazaquistaorepublicado','cazaquistao').str.replace('coreia(dosul)republicada','coreiadosul').str.replace(
                            'russiafederacaoda','russia')

df_trat['país_emissor_do_documento_principal'] = df_trat['país_emissor_do_documento_principal'].str.normalize('NFKD').str.encode('ascii', errors='ignore').str.decode('utf-8').str.replace(
                            ' ','').str.lower().str.replace(',','').str.replace('armeniarepublicada','amermenia').str.replace('chinarepublicapopular','china').str.replace(
                            'congorepublicademocraticado','congo')
df_trat['país_emissor_do_documento_principal'] = df_trat['país_emissor_do_documento_principal'].str.replace('caboverderepublicade','caboverde').str.replace('estoniarepublicada','estonia').str.replace(
                            'georgiarepublicada','georgia').str.replace('croacia(republicada)','croacia').str.replace('hungriarepublicada','hungria').str.replace(
                            'irarepublicaislamicado','ira').str.replace('cazaquistaorepublicado','cazaquistao').str.replace('coreia(dosul)republicada','coreiadosul').str.replace(
                            'russiafederacaoda','russia')

df_trat['pais_de_origem_do_voo'] = df_trat['pais_de_origem_do_voo'].str.normalize('NFKD').str.encode('ascii', errors='ignore').str.decode('utf-8').str.replace(' ','').str.lower().str.replace(
                            ',','').str.replace('armeniarepublicada','amermenia').str.replace('chinarepublicapopular','china').str.replace(
                            'congorepublicademocraticado','congo')
df_trat['pais_de_origem_do_voo']= df_trat['pais_de_origem_do_voo'].str.replace('caboverderepublicade','caboverde').str.replace('estoniarepublicada','estonia').str.replace(
                            'georgiarepublicada','georgia').str.replace('croacia(republicada)','croacia').str.replace('hungriarepublicada','hungria').str.replace(
                            'irarepublicaislamicado','ira').str.replace('cazaquistaorepublicado','cazaquistao').str.replace('coreia(dosul)republicada','coreiadosul').str.replace(
                            'russiafederacaoda','russia')

df_trat['pais_de_embarque_do_viajante'] = df_trat['pais_de_embarque_do_viajante'].str.normalize('NFKD').str.encode('ascii', errors='ignore').str.decode('utf-8').str.replace(
                            ' ','').str.lower().str.replace(',','').str.replace('armeniarepublicada','amermenia').str.replace('chinarepublicapopular','china').str.replace(
                            'congorepublicademocraticado','congo')
df_trat['pais_de_embarque_do_viajante'] = df_trat['pais_de_embarque_do_viajante'].str.replace('caboverderepublicade','caboverde').str.replace('estoniarepublicada','estonia').str.replace(
                            'georgiarepublicada','georgia').str.replace('croacia(republicada)','croacia').str.replace('hungriarepublicada','hungria').str.replace(
                            'irarepublicaislamicado','ira').str.replace('cazaquistaorepublicado','cazaquistao').str.replace('coreia(dosul)republicada','coreiadosul').str.replace(
                            'russiafederacaoda','russia')

df_trat['documento_emissor_vu'] = df_trat['documento_emissor_vu'].str.normalize('NFKD').str.encode('ascii', errors='ignore').str.decode('utf-8').str.replace(
                            ' ','').str.lower().str.replace(',','').str.replace('armeniarepublicada','amermenia').str.replace('chinarepublicapopular','china').str.replace(
                            'congorepublicademocraticado','congo')
df_trat['documento_emissor_vu'] = df_trat['documento_emissor_vu'].str.replace('caboverderepublicade','caboverde').str.replace('estoniarepublicada','estonia').str.replace(
                            'georgiarepublicada','georgia').str.replace('croacia(republicada)','croacia').str.replace('hungriarepublicada','hungria').str.replace(
                            'irarepublicaislamicado','ira').str.replace('cazaquistaorepublicado','cazaquistao').str.replace('coreia(dosul)republicada','coreiadosul').str.replace(
                            'russiafederacaoda','russia')

cod_paises = pd.read_csv('paises_codigos.csv', sep=';')
cod_paises.head(4)

df_trat = pd.merge(df_trat,cod_paises,left_on='nacionalidade',right_on='cod', how='left')
df_trat['nacionalidade_trat'] = df_trat['nacionalidade_trat'].fillna('marrocos')
df_trat = df_trat.drop(columns=['nacionalidade','cod'])
df_trat.head(4)

cod_paises = cod_paises.rename(columns={'nacionalidade_trat':'nacionalidade_vu_trat'})

df_trat['nacionalidade_vu'] = df_trat['nacionalidade_vu'].str.lower()
df_trat = pd.merge(df_trat,cod_paises,left_on='nacionalidade_vu',right_on='cod', how='left')
df_trat['nacionalidade_vu_trat'] = df_trat['nacionalidade_vu_trat'].fillna('brasil') #NA
df_trat = df_trat.drop(columns=['nacionalidade_vu','cod'])
# df_trat['nacionalidade_vu_trat'].head(4)

 CPF x CPNJ - se tiver CPF e não tiver informação no CNPJ, será considerado como 0.

df_trat = df_trat.drop(columns=['cnpj_atividadeeconomicapreferencial_vu'])

# cnpj_qtd_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['cnpj_qtd_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['cnpj_qtd_trat'] = lista_cnpj2

# cnpj_nrestabelecimento_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['cnpj_nrestabelecimento_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['cnpj_nrestabelecimento_trat'] = lista_cnpj2

# situacao_cnpj_ativa_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['situacao_cnpj_ativa_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['situacao_cnpj_ativa_trat'] = lista_cnpj2

# situacao_cnpj_baixada_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['situacao_cnpj_baixada_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['situacao_cnpj_baixada_trat'] = lista_cnpj2

# situacao_cnpj_inapta_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['situacao_cnpj_inapta_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['situacao_cnpj_inapta_trat'] = lista_cnpj2

# situacao_cnpj_nula_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['situacao_cnpj_nula_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['situacao_cnpj_nula_trat'] = lista_cnpj2

# situacao_cnpj_suspensa_trat
lista_cpf,lista_cnpj,lista_cnpj2 = list(df_trat['cpf_trat']),list(df_trat['situacao_cnpj_suspensa_trat']),[]
index = 0
for i in lista_cpf:
    if i == True and np.isnan(lista_cnpj[index]):
        lista_cnpj2.append(0)
    else:
        lista_cnpj2.append(lista_cnpj[index])
    index += 1
df_trat['situacao_cnpj_suspensa_trat'] = lista_cnpj2

 Renda e Patrimônio

df_trat['dirpf_totalrendimentotributavel_vu'] = df_trat['dirpf_totalrendimentotributavel_vu'].apply(lambda x: str(x).replace('.','')).apply(lambda x: str(x).replace(',','.')).apply(lambda x: str(x).replace('R$','')).astype('float64')
df_trat['dirpf_totalrendimentonaotributavel_vu'] = df_trat['dirpf_totalrendimentonaotributavel_vu'].apply(lambda x: str(x).replace('.','')).apply(lambda x: str(x).replace(',','.')).apply(lambda x: str(x).replace('R$','')).astype('float64')
df_trat['dirpf_totalbensdeclarados_vu'] = df_trat['dirpf_totalbensdeclarados_vu'].apply(lambda x: str(x).replace('.','')).apply(lambda x: str(x).replace(',','.')).apply(lambda x: str(x).replace('R$','')).astype('float64')


 Total do rendimento

df_trat['dirpf_totalrendimento_trat'] = df_trat['dirpf_totalrendimentotributavel_vu'] + df_trat['dirpf_totalrendimentonaotributavel_vu']

Forma de Pagamento da passagem

# forma de pagamento
conditions = [
            (df_trat['forma_pgto_pnr'] == '\ Cartao'),
            (df_trat['forma_pgto_pnr'] == '\ Miscelânea'),
            (df_trat['forma_pgto_pnr'] == '\ Dinheiro \ Cartao'),
            (df_trat['forma_pgto_pnr'] == '\ Dinheiro'),  
            (df_trat['forma_pgto_pnr'] == '\ Cartao \ Miscelânea'),
            (df_trat['forma_pgto_pnr'].isna()), 
            ]
df_trat['forma_pgto_pnr_trat'] = np.select(conditions, [1,2,3,4,5,np.nan], default=1)

df_vu = df_trat['cod_vu']
df_vu = df_vu.drop_duplicates()
df_vu.shape

Duração da viagem

df_duracao = pd.read_csv('DURACAO BuscaVU_PUC.csv',sep=';',encoding="latin-1")
df_duracao.columns= ['tipoVoo','ciaAerea','sequencial','numeroVoo','aeroportoOrigem','aeroportoDestino','dataPartida','dataChegada','qtBagagens','localizadorReserva','qtViajantesReserva','tipoViajante','nomeUtilizado','nomeAeroportoOrigem','nomeAeroportoDestino','aeroportoOrigemViajante','aeroportoOrigemLatitude','aeroportoOrigemLongitude','aeroportoOrigemCidade','aeroportoDestinoViajante','aeroportoDestinoLatitude','aeroportoDestinoLongitude','aeroportoDestinoCidade','nomeAeroportoOrigemViajante','nomeAeroportoDestinoViajante','dataNascimento','documentos','apvoNrSq','apvaNrSq','apvjNrSq','assento','duracaoDias','apvjQtPesoBagKg','viagemAnterior','quantidadePesoBagagem','qtdBagagemPesoSaida','qtdBagagemPesoChegada','dataMovimentoNoBrasilFormatada','dataMovimentoNoBrasil','localizadorReservaComQtdViajantes']

def transformar_em_data(dt):
    dia, mes, ano = int(dt[0:2]), int(dt[3:5]), int(dt[6:10])
    data = date(ano, mes, dia)
    return data
    
data_chegada_lista, data_chegada_lista_2 = list(df_duracao['dataChegada'].astype(str)), []
for i in data_chegada_lista:
    data_chegada_lista_2.append(transformar_em_data(i))

df_duracao['data_da_chegada_do_voo'] = data_chegada_lista_2
data_partida_lista, data_partida_lista_2 = list(df_duracao['dataPartida'].astype(str)), []
for i in data_partida_lista:
    data_partida_lista_2.append(transformar_em_data(i))

df_duracao['data_da_partida_do_voo'] = data_partida_lista_2

sequencial = list(df_duracao['sequencial'])
linhas = len(sequencial)
ls_duracao = []
i = 0
while i < linhas-1:
    if df_duracao['sequencial'][i] == df_duracao['sequencial'][i+1]:
        if df_duracao['tipoVoo'][i] == 'E' and df_duracao['tipoVoo'][i+1] == "S":
            ls_duracao.append(str(df_duracao['data_da_chegada_do_voo'][i] - df_duracao['data_da_partida_do_voo'][i+1])) 
        elif df_duracao['tipoVoo'][i] == 'E' and df_duracao['tipoVoo'][i+1] == "E": 
            ls_duracao.append(np.nan) 
        else:
            ls_duracao.append(np.nan) 

    elif df_duracao['sequencial'][i] != df_duracao['sequencial'][i+1]:
        if df_duracao['tipoVoo'][i] == 'E':
            ls_duracao.append(str(-1000))
        else:
            ls_duracao.append(np.nan)

    else:
        ls_duracao.append(np.nan)

    i+=1
ls_duracao.append(np.nan)

df_duracao['duracao_viagem_trat'] = ls_duracao

df_duracao2 = df_duracao.copy()

df_duracao = df_duracao.dropna(axis='rows')

---------------------------------
Merge df_trat + df_duracao

###### Criando código para o merge

def transformar_em_data_chegada_merge(dt):
    dia, mes, ano = str(dt[0:2]), str(dt[3:5]), str(dt[6:10])
    data = dia + mes + ano
    return data

df_duracao['sequencial'] = df_duracao['sequencial'].astype(str)
df_duracao['duracao_viagem_trat'] = df_duracao['duracao_viagem_trat'].astype(str)

lista_data, lista_data_merge = list(df_duracao['dataChegada']), []
for i in lista_data:
    lista_data_merge.append(transformar_em_data_chegada_merge(i)) # transformando a data da chegada
df_duracao['dataChegadaMerge'] = lista_data_merge
df_duracao['vu chegada'] = df_duracao['sequencial'].astype(str) + df_duracao['dataChegadaMerge']
# df_duracao['vu chegada']

df_duracao = df_duracao[['vu chegada','duracao_viagem_trat']]
# df_duracao.head(110)

Merge

df_trat = pd.merge(df_trat,df_duracao, left_on='vu chegada',right_on='vu chegada', how='left')
df_trat.shape

df_trat['duracao_viagem_trat'] = df_trat['duracao_viagem_trat'].str.replace(' days, 0:00:00','').str.replace(' day, 0:00:00','').str.replace('0:00:00','0')
df_trat['duracao_viagem_trat'] = df_trat['duracao_viagem_trat'].astype(float)

df_trat2 = df_trat.copy()
df_trat2 = df_trat2.drop(columns=['cod_vu','data_de_partida_do_voo','data_da_chegada_do_voo2',
                                'cpf_vu','cpfresponsavel_vu','documento_validade_vu','datanascimento_vu',
                                'qtviagens_vu','fiscalizacoes_qt_vu','declaracoes_qt_vu',
                                'cnpj_cod_vu','cnpj_nrestabelecimento_vu',
                                'cnpj_situacao_vu',
                                'doc_numdocumento_hist_vu','doc_paisemissor_hist_vu','origem_lista_vu',
                                'origem_qt_vu','destino_lista_vu','destino_qt_vu','ciaaereahist_vu','numerovoohist_vu','qtviajantesreservahist_vu',
                                'aeroportoorigemviajantehist_vu','aeroportodestinoviajantehist_vu',
                                'data_reserva_pnr','data_reserva_pnr_trat','datanascimento_vu','data_da_chegada_do_voo','viagens_periodo_do_vu_trat','qtd_fisc_descendente_trat',
                                'data_merge_dia','data_merge_mes','data_merge_ano','data_merge','vu chegada','num_viagens_api_subtrair','num_viagens_api_subtrair',
                                'data_de_partida_do_voo_trat', 'origem_ls_trat','origem_ls_trat_2',
                                'qtviajantesreservahist_vu2',
                                'doc_ano_validade_trat','cidade_do_embarque_do_viajante','forma_pgto_pnr' ])
                               

df_trat2.columns

# df_trat2.to_csv('análise_dados_parte1.csv',sep=';')

-----------------
## Análise prévia 

df_analise = df_trat2.copy()

Dados faltantes

df_nan = pd.DataFrame(df_analise.isna().sum())
df_nan.loc[df_nan[0]>0]

Cardinalidade 

features_categoricas = ['aeroporto_de_origem_do_voo', 'país_emissor_do_documento_principal', 'pais_de_origem_do_voo', 'numero_do_voo', 'companhia_aerea', 'codigo_iata_do_local_de_embarque_do_passageiro', 
                        'pais_de_embarque_do_viajante', 'codigo_iata_do_local_de_desembarque_do_viajante', 'documento_emissor_vu', 'nacionalidade_trat', 'nacionalidade_vu_trat']


df_unique = pd.DataFrame(df_analise[features_categoricas].nunique())
df_unique.loc[df_unique[0]>10]
